TC1004/Reto/requirements.txt
# /opt/mqtt-dashboard/requirements.txt
flask
flask-socketio
eventlet
paho-mqtt